Bus activity check
Added functionality to healthcheck to check if there is bus activity and use that to determine health.
If there is no activity on the bus, then the logger is not expected to be doing anything and is still
presumed to be healthy.

Also added some sysout debug flushes to can_logger that assisted me in debugging