x86 fixes (#116)
* Ansible x86 fixes

GRUB default variables were misnamed and systemd command for Network Manager was written as
network-manager instead of NetworkManager

* Clock rounding fix

The new x86 boards seem to give a decimal output for the time, which ksuid does not like. Added a function to round the values.
We do not need sub-millisecond precision for gps location

* x86 preseed file

Currently a seperate file, would like to merge with the toradex version

* Split lattepanda configs into two files

Differences in ssd/nvme config require this for now

* Preseed file fixes

Further fixes for lattepanda preseeds. Specifically moves the boot partition and grub onto the interal mmc allowing booting in non-legacy mode in the alpha and booting at all in the delta

* Docs

* Addition of new files in codeowners folder